+++ test/test_pdfz_syst.cpp
@@ -183,7 +183,7 @@ protected:
 
         pdf_values = new hemi::Array<float>(20, true);
         norm = new hemi::Array<unsigned int>(3, true);
-        params = new hemi::Array<float>(5, true);
+        params = new hemi::Array<double>(5, true);
         params->writeOnlyHostPtr(); // Force memory allocation
 
         evaluator->SetEvalPoints(eval_points);
@@ -209,7 +209,7 @@ protected:
     std::vector<float> eval_points;
     hemi::Array<float> *pdf_values;
     hemi::Array<unsigned int> *norm;
-    hemi::Array<float> *params;
+    hemi::Array<double> *params;
     pdfz::ResolutionScaleSystematic *res;
 };
